medlens/pyproject.toml
[project]
name = "med"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "chromadb>=1.1.0",
    "dotenv>=0.9.9",
    "fastapi==0.109.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-huggingface>=1.2.0",
    "modal>=1.3.2",
    "opencv-python>=4.13.0.92",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "psycopg2-binary>=2.9.11",
    "pydicom>=3.0.1",
    "python-magic-bin>=0.4.14",
    "python-multipart==0.0.6",
    "scikit-image>=0.26.0",
    "sentence-transformers>=5.2.2",
    "torch>=2.10.0",
    "uvicorn[standard]==0.27.0",
]
